aisrch/pyproject.toml
[tool.poetry]
name = "tutoria-aisrch"
version = "0.1.0"
description = ""
authors = ["Sara San Luís Rodríguez, Christian Carballo Lozano"]
readme = "README.md"
keywords = ["tutor.ia", "Azure AI Search"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
azure-identity = "^1.18.0"
azure-mgmt-search = "^9.1.0"
azure-search-documents = "^11.5.1"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
azure-core = "^1.31.0"

[tool.poetry.group.dev.dependencies]
pyclean = "^3.0.0"
ruff = "^0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"